Added smart-tabs-mode.el.
It provides the smart tabs functionality as a minor mode.
See the comments for usage details and examples.

There is one significant difference in behavior compared to the old
code. Before, you had to turn on indent-tabs-mode to use smart tabs.
Now, smart-tabs-mode always uses tabs for indentation, even if
indent-tabs-mode is not on. Rationale: if the user has turned on
smart-tabs-mode, that's a clear sign they want tabs for indentation.